For pull_requests, use the original commit SHA for details.
When a PR is made, github runs CI not on the head of the PR but on the
head of the PR merged with the head of the branch on to which the PR
will merge.  Coveralls should use the details from the head of the PR
branch, not those of the merge that github does.